requirements.txt
requests==2.31.0
Pillow==8.2.0
tensorflow==2.10.0
scikit-learn==0.24.2
numpy==1.21.6
typing-extensions==3.7.4.3
protobuf==3.19.6
